requirements.txt
asgiref==3.5.2
autopep8==1.6.0
crispy-bootstrap5==0.6
Django==4.1.3
django-cors-headers==3.13.0
django-crispy-forms==1.14.0
djangorestframework==3.14.0
javascript==1!1.0.0
psycopg2==2.9.3
pycodestyle==2.8.0
pytz==2022.6
sqlparse==0.4.2
toml==0.10.2
tzdata==2022.1
whitenoise==6.2.0
